Modularized the user functionality of api/index.js
